openstack-api/requirements.txt
apache-libcloud==1.2.1
appdirs==1.4.0
appnope==0.1.0
Babel==2.3.4
cffi==1.8.3
cliff==2.2.0
cmd2==0.6.9
cryptography==1.5.2
debtcollector==1.8.0
decorator==4.0.10
dogpile.cache==0.6.2
futures==3.0.5
idna==2.1
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
iso8601==0.1.11
jsonpatch==1.14
jsonpointer==1.10
jsonschema==2.5.1
keystoneauth1==2.12.1
monotonic==1.2
msgpack-python==0.4.8
munch==2.0.4
netaddr==0.7.18
netifaces==0.10.5
openstacksdk==0.9.8
os-client-config==1.21.1
osc-lib==1.1.0
oslo.config==3.17.0
oslo.i18n==3.9.0
oslo.serialization==2.13.0
oslo.utils==3.16.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
positional==1.1.1
prettytable==0.7.2
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
pyparsing==2.1.10
python-cinderclient==1.9.0
python-designateclient==2.3.0
python-glanceclient==2.5.0
python-heatclient==1.5.0
python-ironicclient==1.7.0
python-keystoneclient==3.5.0
python-magnumclient==2.3.0
python-mistralclient==2.1.1
python-neutronclient==6.0.0
python-novaclient==6.0.0
python-openstackclient==3.2.0
python-swiftclient==3.1.0
python-troveclient==2.5.0
pytz==2016.7
PyYAML==3.12
requests==2.11.1
requestsexceptions==1.1.3
rfc3986==0.4.1
shade==1.12.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
stevedore==1.17.1
traitlets==4.3.1
warlock==1.2.0
wcwidth==0.1.7
wrapt==1.10.8
